requirements.txt
streamlit==1.32.2
openai==1.3.0
google-generativeai==0.3.2
requests>=2.27.0
beautifulsoup4==4.9.3
python-docx==0.8.11
PyPDF2==3.0.1
python-dotenv==0.19.0
PyYAML==6.0.1
